--- food-pick/foodNameBatch/.aws-sam/build.toml
@@ -1,15 +1,0 @@
-# This file is auto generated by SAM CLI build command
-
-[function_build_definitions.b9d7cae9-389d-4104-92e8-75313ea653c3]
-codeuri = "/Users/mac/project/batch_system/food-pick/foodNameBatch/src"
-runtime = "provided.al2023"
-architecture = "x86_64"
-handler = "bootstrap"
-manifest_hash = ""
-packagetype = "Zip"
-functions = ["FoodNameBatchFunction"]
-
-[function_build_definitions.b9d7cae9-389d-4104-92e8-75313ea653c3.metadata]
-BuildMethod = "go1.x"
-
-[layer_build_definitions]
