+++ food-pick/foodPick/.aws-sam/build.toml
@@ -1,7 +1,7 @@
 # This file is auto generated by SAM CLI build command
 
-[function_build_definitions.3c0f37e6-9336-4f21-a1ee-e5aef1812838]
-codeuri = "/Users/mac/project/batch_system/food-pick/foodPick/src"
+[function_build_definitions.87cefe27-58b6-49c8-989a-90c366713987]
+codeuri = "/Users/mac/project/batch_system/food-pick/foodPick/src/users"
 runtime = "provided.al2023"
 architecture = "x86_64"
 handler = "bootstrap"
