+++ food-pick/foodPick/deploy_dev.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+sam build -t ./template.yaml && sam deploy --config-env apne2-dev --no-progressbar --no-confirm-changeset
