+++ food-pick/foodPick/local_test.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+sam local invoke GetUserFunction --event events/event.json
