+++ food-pick/foodPick/src/common/env.go
@@ -1,4 +1,4 @@
-package utils
+package common
 
 import (
 	"fmt"
