+++ food-pick/foodPick/src/common/error.go
@@ -1,4 +1,4 @@
-package utils
+package common
 
 import (
 	"context"
