+++ food-pick/foodPick/src/common/jwt.go
@@ -1,4 +1,4 @@
-package utils
+package common
 
 import (
 	"crypto/aes"
