+++ food-pick/foodPick/src/common/logging.go
@@ -1,4 +1,4 @@
-package utils
+package common
 
 import (
 	"encoding/json"
