+++ food-pick/foodPick/src/common/validator.go
@@ -1,4 +1,4 @@
-package utils
+package common
 
 import (
 	"context"
