+++ food-pick/foodPick/src/users/.aws-sam/build.toml
@@ -1,0 +1,12 @@
+# This file is auto generated by SAM CLI build command
+
+[function_build_definitions.1a602485-d348-4b72-8ad2-5b45395cb0a1]
+codeuri = "/Users/mac/project/batch_system/food-pick/foodPick/src/users"
+runtime = "provided.al2023"
+architecture = "x86_64"
+handler = "bootstrap"
+manifest_hash = ""
+packagetype = "Zip"
+functions = ["GetUserFunction"]
+
+[layer_build_definitions]
