+++ food-pick/foodPick/src/users/model/interface/IUserRepository.go
@@ -2,9 +2,8 @@ package _interface
 
 import (
 	"context"
-	"main/common/db/mysql"
 )
 
 type IGetUserRepository interface {
-	FindOneUser(ctx context.Context, uID uint) (*mysql.Users, error)
+	FindOneUser(ctx context.Context, uID uint) error
 }
