+++ food-pick/foodPick/src/users/repository/getUser.go
@@ -2,8 +2,7 @@ package repository
 
 import (
 	"context"
-	"main/common/db/mysql"
-	_interface "main/users/model/interface"
+	_interface "main/model/interface"
 
 	"gorm.io/gorm"
 )
@@ -11,9 +10,7 @@ import (
 func NewGetUserRepository(gormDB *gorm.DB) _interface.IGetUserRepository {
 	return &GetUserRepository{GormDB: gormDB}
 }
-func (d *GetUserRepository) FindOneUser(ctx context.Context, uID uint) (*mysql.Users, error) {
+func (d *GetUserRepository) FindOneUser(ctx context.Context, uID uint) error{
 
-	user := mysql.Users{}
-
-	return &user, nil
+	return nil
 }
